backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
beautifulsoup4
requests